modularize Cascader, Checkbox and DatePicker
