+++ src/cascader/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/cascader/style/css.js"
