+++ src/date-picker/style/css.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/date-picker/style/css.js"
