+++ src/date-picker/style/index.tsx
@@ -1,0 +1,1 @@
+import "antd/lib/date-picker/style"
