+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from "./DatePicker";
+export * from "./date-picker";
 export * from "./TimePicker";
 export * from "./FormikDebug";
 export * from "./Input";
@@ -14,7 +14,7 @@ export * from "./Radio";
 export * from "./auto-complete";
 export * from "./Rate";
 export * from "./Slider";
-export * from "./Cascader";
+export * from "./cascader";
 export * from "./Form";
 export * from "./TreeSelect";
 export * from "./Mention";
